feat(website): finalize deploy section as V5 with accurate copy
Picks V5 (two columns with bullets and vertical OR rail) as the
final layout. Inlines it into CloudProviders.tsx and removes the
exploratory variants file.

Copy verified against docs/hosting/introduction.mdx and
docs/cli-reference/open-and-connect.mdx — names Browserbase/Kernel
as supported managed providers and points self-host at the
Dockerfile + Cloud Run / ECS guides.
